rename `member.py` to `guild.py`
Member is actually a `guild member`
